separate examples
