+++ frontend/docs/pages/home/run-with-results.mdx
@@ -7,7 +7,7 @@ export const SimpleTs = {
 };
 
 export const SimplePy = {
-  path: "examples/simple/trigger.py",
+  path: "examples/child/trigger.py",
 };
 
 export const SimpleGo = {
