+++ sdks/python/examples/child/bulk.py
